Refactor data handling and improve readability by changing data types to size_t, adding a Range class for better range management, and enhancing serial output formatting in firmware. Also, update transformer calculations and constants for clarity and accuracy.

+++ firmware/data.h
@@ -1,28 +1,25 @@
-constexpr uint16_t DATA_SIZE = pow(2, 8);
-constexpr uint16_t DATA_MAX_IDX = DATA_SIZE - 1;
-constexpr uint16_t DATA_SIZE_QUATER = DATA_SIZE / 4;
+constexpr size_t DATA_SIZE = pow(2, 8);
+constexpr size_t DATA_MAX_IDX = DATA_SIZE - 1;
+constexpr size_t DATA_SIZE_QUATER = DATA_SIZE / 4;
 
 uint32_t data[DATA_SIZE] = {0};
-uint16_t head = 0;
+size_t head = 0;
 
 void data_push(uint16_t value)
 {
-//    Serial.print('h');
-//    Serial.println(head);
 //    data[head] = value;
     data[head] = (uint32_t)value * (uint32_t)value;
     head = (head + 1) & (DATA_MAX_IDX);
-
 }
 
 uint16_t data_rms() {
   uint16_t count = 0;
-  uint64_t sum = 0;
-  for (uint16_t i = 0; i < DATA_SIZE; i++) {
+  uint32_t sum = 0;
+  for (size_t i = 0; i < DATA_SIZE; i++) {
     if (data[i] > 0) {
       count++;
 //      sum += (uint32_t)data[i] * (uint32_t)data[i];
-      sum += (uint64_t)data[i];
+      sum += data[i];
     }
   }
   return TEST || count > DATA_SIZE_QUATER
@@ -33,7 +30,7 @@ uint16_t data_rms() {
 
 uint16_t data_max() {
   uint32_t max = data[0];
-  for (uint16_t i = 1; i < DATA_SIZE; i++) {
+  for (size_t i = 1; i < DATA_SIZE; i++) {
     if (max < data[i]) {
       max = data[i];
     }
